Initial design (#8)
* initial designs

This displays a list of movies with the discover endpoint from the movie db. Please note the movie trailers are currently static and only show one movie trailer.

* Initial Design

Currently it will fetch a list of movies. Then you can click a button to load more. You can also view movie trailers in a modal.

* Initial Design Modifications

Totally revamped the CSS to use BEM and CSS nesting. Updated all the templates that used the old CSS class names. Renamed the modal function name to something more understandable. Deleted old commented out code.

* Initial Design Additional Modifications

I separated the formatMovieTrailers into several helper functions. Then I renamed it to more align with when we use it for TV show trailers. It is now trailersTemplate.

* Initial Design

Renamed the helpers folder to partials and updated all the dependencies or imports.